modify makefile and run_tests fix compiler warnings

+++ tests/signals/kill-child.c
@@ -82,7 +82,7 @@ void kill_child(int signum)
 
 int main()
 {
-	for (int i = 1; i <  sizeof(signals_list)/sizeof(signals_list[0]); i++)
+	for (long unsigned int i = 1; i <  sizeof(signals_list)/sizeof(signals_list[0]); i++)
 	{
 		int sig = signals_list[i].signal;
 		if (sig == SIGKILL || sig == SIGSTOP)
